Remove obsolete del_objects.py
