+++ collective/deletepermission/del_object_dexterity.py
@@ -1,5 +1,23 @@
+from AccessControl import Unauthorized
+from AccessControl import getSecurityManager
 from plone.dexterity.content import Container
-from collective.deletepermission.del_object import protect_del_objects
+
+
+def protect_del_objects(self, ids=None):
+    sm = getSecurityManager()
+    if not sm.checkPermission('Delete objects', self):
+        raise Unauthorized(
+            "Do not have permissions to remove this object")
+
+    if ids is None:
+        ids = []
+    if isinstance(ids, str):
+        ids = [ids]
+    for id_ in ids:
+        item = self._getOb(id_)
+        if not sm.checkPermission("Delete portal content", item):
+            raise Unauthorized(
+                "Do not have permissions to remove this object")
 
 
 def manage_delObjects(self, ids=None, REQUEST=None):
